feat : 카드 정보 추가
